refactor: extract collapsible height animation into shared hook
- Create useCollapsibleHeight hook for smooth expand/collapse animations
- Update CollapsibleSection in AgentPnL.tsx to use the hook
- Update Collapsible in AgentSettingsSidebar.tsx to use the hook
- Eliminates duplicate animation logic across components
